+++ tests/fixtures/ephemeral/D.py
@@ -1,0 +1,5 @@
+def configure(context):
+    context.stage("tests.fixtures.ephemeral.C")
+
+def execute(context):
+    pass